refactor(projen.project.typescript): extract `builders` module
