+++ packages/projen/project/typescript/src/index.ts
@@ -2,3 +2,4 @@ export * from './typescript'
 export * from './typescript-project-options'
 export * from './typescript-config-options'
 export * from './typescript-compiler-options'
+export * as builders from './builders'
